requirements.txt
pandas
numpy
PyYAML
requests
pydantic
python-dotenv
nuitka
zstandard
